Change the Code of Conduct
This commit replaces references to Software Carpentry Code of Conduct
(with whom we are not affiliated with them) with a "standard"
Contributor Covenant Code of Conduct, following the discussion in #3.

TODO: a placeholder with a contact method for reporting violoations
should be added before the first workshop. Ideally this would be
the tutor plus another person who is the tutor's superior, but not
directly involved.